moving benchmarking components from view folder to app/benchmarking
