+++ app/benchmarking/components/benchmarkReportViewer.tsx
@@ -1,5 +1,6 @@
 'use client';
 import { useEffect, useState } from 'react';
+import { downloadHtmlReport } from '@/app/benchmarking/utils/reportDownloader';
 import { Button, ButtonType } from '@/app/components/button';
 import { SelectInput } from '@/app/components/selectInput';
 import { useGetBenchmarksResultQuery } from '@/app/services/benchmark-api-service';
@@ -8,7 +9,6 @@ import { colors } from '@/app/views/shared-components/customColors';
 import { LoadingAnimation } from '@/app/views/shared-components/loadingAnimation';
 import { MainSectionSurface } from '@/app/views/shared-components/mainSectionSurface/mainSectionSurface';
 import { BenchmarkReport } from './benchmarkReport';
-import { downloadHtmlReport } from './utils/reportDownloader';
 
 function BenchmarkReportViewer() {
   const [selectedEndpointId, setSelectedEndpointId] = useState('');
