+++ app/benchmarking/components/gradeColors.ts
@@ -1,4 +1,4 @@
-import { GradingColors } from '@/app/views/benchmarking/types/benchmarkReportTypes';
+import { GradingColors } from '@/app/benchmarking/types/benchmarkReportTypes';
 
 export const gradeColorsMoonshot: GradingColors = {
   A: '#25A167',
