+++ app/benchmarking/cookbooks/cookbooksMain.tsx
@@ -1,11 +1,11 @@
 'use client';
 import { useRouter } from 'next/navigation';
 import React from 'react';
+import BenchmarkRunForm from '@/app/benchmarking/components/benchmarkRunForm';
 import { CookbooksViewList } from '@/app/benchmarking/cookbooks/cookbooksViewList';
 import { Icon, IconName } from '@/app/components/IconSVG';
 import SimpleStepsIndicator from '@/app/components/simpleStepsIndicator';
 import { NewEndpointForm } from '@/app/endpoints/(edit)/newEndpointForm';
-import BenchmarkRunForm from '@/app/views/benchmarking/benchmarkRunForm';
 import { EndpointSelectVew } from '@/app/views/models-management/endpointsSelector';
 import { colors } from '@/app/views/shared-components/customColors';
 import { MainSectionSurface } from '@/app/views/shared-components/mainSectionSurface/mainSectionSurface';
