+++ app/benchmarking/report/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { BenchmarkReportViewer } from '@/app/views/benchmarking/benchmarkReportViewer';
+import { BenchmarkReportViewer } from '@/app/benchmarking/components/benchmarkReportViewer';
 
 export default function BenchmarkingReportPage() {
   return <BenchmarkReportViewer />;
