+++ app/benchmarking/session/new/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { BenchmarkNewSessionFlow } from '@/app/views/benchmarking/benchmarkNewSessionFlow';
+import { BenchmarkNewSessionFlow } from '@/app/benchmarking/components/benchmarkNewSessionFlow';
 
 export default function BenchmarkNewSessionFlowPage() {
   return <BenchmarkNewSessionFlow />;
