+++ app/redteaming/(fullscreen)/components/viewBookmarksModal.tsx
@@ -5,9 +5,9 @@ import { Icon, IconName } from '@/app/components/IconSVG';
 import { TextInput } from '@/app/components/textInput';
 import { formatDate } from '@/app/lib/date-utils';
 import { colors } from '@/app/views/shared-components/customColors';
+import { LoadingAnimation } from '@/app/views/shared-components/loadingAnimation';
 import { Modal } from '@/app/views/shared-components/modal/modal';
 import { ColorCodedTemplateString } from './color-coded-template';
-import { LoadingAnimation } from '@/app/views/shared-components/loadingAnimation';
 
 type ViewBookmarksModalProps = {
   onCloseIconClick: () => void;
