+++ app/services/benchmark-api-service.ts
@@ -1,5 +1,5 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import { CookbooksBenchmarkResult } from '@/app/views/benchmarking/types/benchmarkReportTypes';
+import { CookbooksBenchmarkResult } from '@/app/benchmarking/types/benchmarkReportTypes';
 import {
   proxyPathBenchmarksExec,
   proxyPathBenchmarksGetResults,
